pyproject.toml
# Poetry-specific settings.
[project]
name = "multi-agent-ai-system"
version = "0.1.0"
description = "Multi-Agent AI system"
requires-python = ">=3.12, <3.13"
readme = "README.md"
authors = []

dependencies = [
    "pydantic~=2.12.5",
    "python-dateutil>=2.9.0,<3",
    "pytest-dotenv>=0.5.2,<0.6",

    "mlflow~=3.7.0",
    "google-adk~=1.20.0"

]

# Development dependencies of the project.
[dependency-groups]
dev = [
    "pytest~=9.0.2",
    "pytest-mock~=3.15.1",
    "pytest-dotenv~=0.5.2",
    "pre-commit~=4.5.0",
    "ruff~=0.14.8",
    "pyright~=1.1.407",
    "pytest-asyncio~=1.3.0",
    "pytest-cov~=7.0.0",
    "setuptools~=80.9.0",
    "nbstripout~=0.8.2",

]


[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Ruff section configures the Ruff linter and formatter.
[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
typeCheckingMode = "basic"

[tool.ruff]
exclude = ["notebooks"]
line-length = 150
